refactor: move array into StoredHooks.tsx
